requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database dependencies
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Data processing and simulation
numpy==1.26.2
pandas==2.1.3
scipy==1.11.4

# Date/time handling
python-dateutil==2.8.2
pytz==2023.3

# API documentation
python-multipart==0.0.6

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Data validation
email-validator==2.1.0

# CORS support
python-jose[cryptography]==3.3.0

# Monitoring and logging
structlog==23.2.0

# Type checking (optional, for development)
mypy==1.7.1
types-python-dateutil==2.8.19.14